requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.11.16
requests==2.31.0
flask==2.3.3